refactor: split npcs to different files and fix Doge stop trigger
